model/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0

# Data Processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
pyarrow>=12.0.0

# Machine Learning & Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Experiment Tracking & Versioning
mlflow>=2.5.0
dvc>=3.0.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Text Processing
regex>=2023.6.3
emoji>=2.8.0

# Optional: For ONNX export
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional: For SHAP interpretability (Phase 3)
shap>=0.42.0

# Development Tools
ipykernel>=6.25.0
ipywidgets>=8.1.0
jupyter>=1.0.0
